requirements.txt
argon2-cffi==19.2.0
automium==0.2.6
bcrypt==3.1.7
bootstrap4==0.1.0
certifi==2019.9.11
cffi==1.13.2
Click==7.0
Django==2.2.5
django-crispy-forms==1.8.1
Faker==2.0.3
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
legacy==0.1.6
MarkupSafe==1.1.1
Pillow==7.0.0
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
Werkzeug==0.16.0
whitenoise==5.0.1
